rename tours.js to useTourStore
avoid confusion with other stores
