+++ resources/camino-creator/components/Stage/stages/Quiz.vue
@@ -56,7 +56,7 @@
 </template>
 
 <script setup>
-import { useTourStore } from "../../../stores/tours";
+import { useTourStore } from "@creator/stores/useTourStore";
 import LanguageText from "../../LanguageText.vue";
 import { createMultilingualText } from "./stageFactory";
